Move complex utilities out of Half.h
